ADD inMemoryRepoTest Take() & GetByIds()
Wycofałem EfRepoTests (bo i tak były "złe"), a zająłęm się
InMemoryRepoTests, ponieważ chciałem iść zgodnie z kolejnością z
ćwiczeń.
