+++ kod/UamTTA/UamTTA.Tests/EfRepositoryTests.cs
@@ -141,102 +141,108 @@ namespace UamTTA.Tests
             CollectionAssert.AllItemsAreUnique(result);
         }
 
-        [Test]
-        public void Take_Should_Throw_Exception_When_Repository_Is_Empty()
-        {
-            //Arrange
-            int count = 5;
+        //[Test]
+        //public void Take_Should_Throw_Exception_When_Repository_Is_Empty()
+        //{
+        //    //Arrange
+        //    int count = 5;
 
-            //Act & Assert
-            Assert.Throws<ArgumentException>(() => _sut.Take(count));
-        }
+        //    //Act & Assert
+        //    Assert.Throws<ArgumentException>(() => _sut.Take(count));
+        //}
 
-        [Test]
-        public void Take_Should_Throw_Exception_When_There_Is_Less_Objects_Than_Count()
-        {
-            //Arrange
-            int count = 5;
+        //[Test]
+        //public void Take_Should_Throw_Exception_When_There_Is_Less_Objects_Than_Count()
+        //{
+        //    //Arrange
+        //    int count = 5;
 
-            var model1 = new Account { Id = null, Balance = 10, Name = "Bla" };
-            var model2 = new Account { Id = null, Balance = 12, Name = "BlaBla" };
+        //    var model1 = new Account { Id = null, Balance = 10, Name = "Bla" };
+        //    var model2 = new Account { Id = null, Balance = 12, Name = "BlaBla" };
 
-            _sut.Persist(model1);
-            _sut.Persist(model2);
+        //    _sut.Persist(model1);
+        //    _sut.Persist(model2);
 
-            //Act & Assert
-            Assert.Throws<ArgumentException>(() => _sut.Take(count));
-        }
+        //    //Act & Assert
+        //    Assert.Throws<ArgumentException>(() => _sut.Take(count));
+        //}
 
-        [Test]
-        public void Take_Should_Return_First_Count_Elements_From_Repository()
-        {
-            //Arrange
-            int count = 2;
+        //[Test]
+        //public void Take_Should_Return_First_Count_Elements_From_Repository()
+        //{
+        //    //Arrange
+        //    int count = 2;
 
-            var model1 = new Account { Id = null, Balance = 10, Name = "Bla" };
-            var model2 = new Account { Id = null, Balance = 12, Name = "BlaBla" };
-            var model3 = new Account { Id = null, Balance = 40, Name = "BlaBlaBla" };
+        //    var model1 = new Account { Id = null, Balance = 10, Name = "Bla" };
+        //    var model2 = new Account { Id = null, Balance = 12, Name = "BlaBla" };
+        //    var model3 = new Account { Id = null, Balance = 40, Name = "BlaBlaBla" };
 
-            _sut.Persist(model1);
-            _sut.Persist(model2);
-            _sut.Persist(model3);
+        //    _sut.Persist(model1);
+        //    _sut.Persist(model2);
+        //    _sut.Persist(model3);
 
-            //Act
-            var result = _sut.Take(count);
+        //    //Act
+        //    var result = _sut.Take(count);
 
-            //Assert
-            Assert.That(result, Is.Not.Null);
-            Assert.That(result.Count(), Is.EqualTo(count));
-        }
+        //    //Assert
+        //    //Assert.That(result, Is.Not.Null);
+        //    Assert.That(result.Count(), Is.EqualTo(count));
+        //    CollectionAssert.Contains(result, _sut.FindById(1));
 
-        [Test]
-        public void GetByIds_Should_Return_Null_When_There_Is_No_Objects_From_Given_Id_Set()
-        {
-            int[] ids = new int[] { 1, 2, 3 };
+        //    foreach (var account in result)
+        //    {
+        //        Assert.That(account, Is.Not.Null);
+        //    }
+        //}
 
-            Account[] result = (Account[])_sut.GetByIds(ids);
+        //[Test]
+        //public void GetByIds_Should_Return_Null_When_There_Is_No_Objects_From_Given_Id_Set()
+        //{
+        //    int[] ids = new int[] { 1, 2, 3 };
 
-            Assert.That(result, Is.Empty);
-            Assert.That(result, Is.Null);
-        }
+        //    var result = _sut.GetByIds(ids);
 
-        [Test]
-        public void GetByIds_Should_Return_All_Objects_From_Given_Id_Set()
-        {
-            var model1 = new Account { Id = null, Balance = 10, Name = "Bla" };
-            var model2 = new Account { Id = null, Balance = 12, Name = "BlaBla" };
+        //    Assert.That(result, Is.Null);
+        //}
 
-            var persisted1 = _sut.Persist(model1);
-            var persisted2 = _sut.Persist(model2);
+        //[Test]
+        //public void GetByIds_Should_Return_All_Objects_From_Given_Id_Set()
+        //{
+        //    var model1 = new Account { Id = null, Balance = 10, Name = "Bla" };
+        //    var model2 = new Account { Id = null, Balance = 12, Name = "BlaBla" };
 
-            int[] ids = new int[] { persisted1.Id.Value, persisted2.Id.Value };
+        //    var persisted1 = _sut.Persist(model1);
+        //    var persisted2 = _sut.Persist(model2);
 
-            Account[] result = (Account[])_sut.GetByIds(ids);
+        //    int[] ids = new int[] { persisted1.Id.Value, persisted2.Id.Value };
 
-            Assert.That(result[0], Is.Not.Null);
-            Assert.That(result[1], Is.Not.Null);
-            Assert.That(result[0].Name, Is.EqualTo("Bla"));
-            Assert.That(result[1].Balance, Is.EqualTo(12));
-        }
+        //    var result = _sut.GetByIds(ids);
 
-        [Test]
-        public void GetByIds_Should_Ommit_When_There_Is_No_Object_With_Given_Id()
-        {
-            var model1 = new Account { Id = null, Balance = 10, Name = "Bla" };
-            var model2 = new Account { Id = null, Balance = 12, Name = "BlaBla" };
+        //    foreach (Account account in result)
+        //    {
+        //        Assert.That(account, Is.Not.Null);
+        //        Assert.That(account.Name, (account.Id.Value == 0) ? Is.EqualTo("Bla") : Is.EqualTo("BlaBla"));
+        //    }
+        //}
 
-            var persisted1 = _sut.Persist(model1);
-            var persisted2 = _sut.Persist(model2);
+        //[Test]
+        //public void GetByIds_Should_Ommit_When_There_Is_No_Object_With_Given_Id()
+        //{
+        //    var model1 = new Account { Id = null, Balance = 10, Name = "Bla" };
+        //    var model2 = new Account { Id = null, Balance = 12, Name = "BlaBla" };
 
-            int[] ids = new int[] { persisted1.Id.Value, 500, 700, persisted2.Id.Value };
+        //    var persisted1 = _sut.Persist(model1);
+        //    var persisted2 = _sut.Persist(model2);
 
-            Account[] result = (Account[])_sut.GetByIds(ids);
+        //    int[] ids = new int[] { persisted1.Id.Value, 500, 700, persisted2.Id.Value };
 
-            Assert.That(result[0], Is.Not.Null);
-            Assert.That(result[1], Is.Not.Null);
-            Assert.That(result[0].Name, Is.EqualTo("Bla"));
-            Assert.That(result[1].Balance, Is.EqualTo(12));
-            Assert.That(result[2], Is.Null);
-        }
+        //    var result = _sut.GetByIds(ids);
+
+        //    Assert.That(result.Count(), Is.EqualTo(2));
+        //    foreach (Account account in result)
+        //    {
+        //        Assert.That(account, Is.Not.Null);
+        //    }
+        //}
     }
 }
